Added randomNumber + method TriFork and tabview NF

+++ functions_basic.py
@@ -2,6 +2,7 @@
 # This file contains the all the functions that are used for GUI. #
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
+import random
 import dotenv, os
 import customtkinter as ctk
 
@@ -10,8 +11,19 @@ def dotENVHelper(dialog, subject):
     dotenv.set_key(dotenv.find_dotenv(), subject, answer)
 
 def dotENVDelete():
-    os.remove(".env")
-    open(".env", "w").close()
+    # Create a random number between 1 and 100
+    randomNumber = random.randint(1, 100)
+    dialog = ctk.CTkInputDialog(title="Delete .env", text=f"Are you sure you want to delete the .env file? \nType {randomNumber} to confirm.")
+    try:
+        answer = int(dialog.get_input())
+    except ValueError:
+        return
+
+    if answer == randomNumber:
+        os.remove(".env")
+        open(".env", "w").close()
+    elif answer == "N":
+        return
 
 def dotENVDiscordBot():
     dialog = ctk.CTkInputDialog(title="Discord Token", text="Enter the relevant details")
@@ -20,3 +32,11 @@ def dotENVDiscordBot():
 def dotENVOpenAI():
     dialog = ctk.CTkInputDialog(title="OpenAI Token", text="Enter the relevant details")
     dotENVHelper(dialog, "OPENAI_TOKEN")
+
+def dotENVTriFork():
+    dialog = ctk.CTkInputDialog(title="First Set", text="Enter the relevant details")
+    dotENVHelper(dialog, "TRIFORK_1")
+    dialog = ctk.CTkInputDialog(title="Second Set", text="Enter the relevant details")
+    dotENVHelper(dialog, "TRIFORK_2")
+    dialog = ctk.CTkInputDialog(title="Owner Confirmation", text="Enter the relevant details")
+    dotENVHelper(dialog, "TRIFORK_OWNER")
